+++ src/hooks/index.ts
@@ -1,1 +1,1 @@
-export * from './useUsers';
+export * from '../references/useUsersRef';
